+++ flag/service/alertmanager/alertmanager.go
@@ -1,11 +1,10 @@
 package alertmanager
 
 type Alertmanager struct {
-	Address    string
-	BaseDomain string
-	LogLevel   string
-	Storage    AlertmanagerStorage
-	Version    string
+	Address  string
+	LogLevel string
+	Storage  AlertmanagerStorage
+	Version  string
 }
 
 type AlertmanagerStorage struct {