+++ service/controller/resource/monitoring/prometheus/resource.go
@@ -249,7 +249,7 @@ func toPrometheus(v interface{}, config Config) (metav1.Object, error) {
 			},
 			RuleNamespaceSelector: &metav1.LabelSelector{
 				MatchLabels: map[string]string{
-					"name": key.NamespaceMonitoring(cluster),
+					"name": key.NamespaceMonitoring(),
 				},
 			},
 			PriorityClassName: "prometheus",