+++ CHANGELOG.md
@@ -7,10 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added Alertmanager managed by Prometheus Operator.
+- Added Alertmanager ingress.
+
 ### Changed
 
 - Lower Prometheus disk space alert from 10% to 5%.
 - Change severity of `ChartOperatorDown` alert to notify.
+- Merge all provider certificate.management-cluster.rules into one prometheus rule.
+
+### Fixed 
+
+- Fix service name in ingress.
 
 ## [1.40.0] - 2021-06-14